+++ src/main/java/com/example9/domain/Order.java
@@ -12,7 +12,7 @@ import java.util.List;
  */
 public class Order {
 	/** 主キー */
-	private Integer id;
+	private Long id;
 	/** ユーザid */
 	private Integer userId;
 	/** 状態 */
@@ -24,7 +24,7 @@ public class Order {
 	/** 宛先氏名 */
 	private String destinationName;
 	/** 宛先Eメール */
-	private String destinationEmail; 
+	private String destinationEmail;
 	/** 宛先郵便番号 */
 	private String destinationZipcode;
 	/** 宛先住所 */
@@ -38,92 +38,120 @@ public class Order {
 	/** ユーザ */
 	private User user;
 	/** 注文商品リスト */
-	private List <OrderItem> orderItemList;
-	
-	public Integer getId() {
+	private List<OrderItem> orderItemList;
+
+	public Long getId() {
 		return id;
 	}
-	public void setId(Integer id) {
+
+	public void setId(Long id) {
 		this.id = id;
 	}
+
 	public Integer getUserId() {
 		return userId;
 	}
+
 	public void setUserId(Integer userId) {
 		this.userId = userId;
 	}
+
 	public Integer getStatus() {
 		return status;
 	}
+
 	public void setStatus(Integer status) {
 		this.status = status;
 	}
+
 	public Integer getTotalPrice() {
 		return totalPrice;
 	}
+
 	public void setTotalPrice(Integer totalPrice) {
 		this.totalPrice = totalPrice;
 	}
+
 	public Date getOrderDate() {
 		return orderDate;
 	}
+
 	public void setOrderDate(Date orderDate) {
 		this.orderDate = orderDate;
 	}
+
 	public String getDestinationName() {
 		return destinationName;
 	}
+
 	public void setDestinationName(String destinationName) {
 		this.destinationName = destinationName;
 	}
+
 	public String getDestinationEmail() {
 		return destinationEmail;
 	}
+
 	public void setDestinationEmail(String destinationEmail) {
 		this.destinationEmail = destinationEmail;
 	}
+
 	public String getDestinationZipcode() {
 		return destinationZipcode;
 	}
+
 	public void setDestinationZipcode(String destinationZipcode) {
 		this.destinationZipcode = destinationZipcode;
 	}
+
 	public String getDestinationAddress() {
 		return destinationAddress;
 	}
+
 	public void setDestinationAddress(String destinationAddress) {
 		this.destinationAddress = destinationAddress;
 	}
+
 	public String getDestinationTel() {
 		return destinationTel;
 	}
+
 	public void setDestinationTel(String destinationTel) {
 		this.destinationTel = destinationTel;
 	}
+
 	public Timestamp getDeliveryTime() {
 		return deliveryTime;
 	}
+
 	public void setDeliveryTime(Timestamp deliveryTime) {
 		this.deliveryTime = deliveryTime;
 	}
+
 	public Integer getPaymentMethod() {
 		return paymentMethod;
 	}
+
 	public void setPaymentMethod(Integer paymentMethod) {
 		this.paymentMethod = paymentMethod;
 	}
+
 	public User getUser() {
 		return user;
 	}
+
 	public void setUser(User user) {
 		this.user = user;
 	}
+
 	public List<OrderItem> getOrderItemList() {
 		return orderItemList;
 	}
+
 	public void setOrderItemList(List<OrderItem> orderItemList) {
 		this.orderItemList = orderItemList;
 	}
+
 	@Override
 	public String toString() {
 		return "Order [id=" + id + ", userId=" + userId + ", status=" + status + ", totalPrice=" + totalPrice
@@ -131,22 +159,22 @@ public class Order {
 				+ destinationEmail + ", destinationZipcode=" + destinationZipcode + ", destinationAddress="
 				+ destinationAddress + ", destinationTel=" + destinationTel + ", deliveryTime=" + deliveryTime
 				+ ", paymentMethod=" + paymentMethod + ", user=" + user + ", orderItemList=" + orderItemList + "]";
-	} 
+	}
 
 	/**
 	 * 合計金額の税額を算出.
+	 * 
 	 * @return 税額
 	 */
 	public int getTax() {
 		int tax = getCalcTotalPrice() * 10 / 100;
 		return tax;
 	}
-	
-	
-	
+
 	/**
 	 * 合計金額（税抜き）を算出.
-	 * @return　合計金額（税抜き）
+	 * 
+	 * @return 合計金額（税抜き）
 	 */
 	public int getCalcTotalPrice() {
 		int totalPrice = 0;
